Add new / and /config/verify pages
