--- client/src/route/config/Developer.tsx
@@ -1,5 +1,0 @@
-import * as React from 'react';
-
-const Developer = () => <></>;
-
-export default Developer;
