--- client/src/route/config/User.tsx
@@ -1,5 +1,0 @@
-import * as React from 'react';
-
-const User = () => <></>;
-
-export default User;
