+++ client/src/routes.tsx
@@ -3,27 +3,36 @@ import WorkBench from 'route/workbench/Workbench';
 import LayoutPublic from 'page/LayoutPublic';
 import PrivateRoute from 'route/auth/PrivateRoute';
 import LibraryPreview from 'preview/route/library/LibraryPreview';
-import VerifyConfig from 'route/config/Verify';
+// import VerifyConfig from 'route/config/Verify';
 import Library from './route/library/Library';
 import DigitalTwins from './route/digitaltwins/DigitalTwins';
 import DigitalTwinsPreview from './preview/route/digitaltwins/DigitalTwinsPreview';
 import SignIn from './route/auth/Signin';
 import Account from './route/auth/Account';
+import Config from './route/config/Config';
 
 export const routes = [
   {
     path: '/',
     element: (
-      <LayoutPublic>
-        <SignIn />
+      <LayoutPublic containerMaxWidth="md">
+        <Config role="user" />
       </LayoutPublic>
     ),
   },
   {
-    path: 'verify',
+    path: 'config/verify',
     element: (
       <LayoutPublic containerMaxWidth="md">
-        <VerifyConfig />
+        <Config role="developer" />
+      </LayoutPublic>
+    ),
+  },
+  {
+    path: 'signin',
+    element: (
+      <LayoutPublic>
+        <SignIn />
       </LayoutPublic>
     ),
   },
