+++ client/test/integration/Routes/Signin.test.tsx
@@ -1,6 +1,6 @@
 import { screen, act } from '@testing-library/react';
 import { setupIntegrationTest } from 'test/integration/integration.testUtil';
-import { getValidationResults } from 'route/config/Verify'; // Globally mocked
+import { getValidationResults } from 'route/config/Verification'; // Globally mocked
 import { testPublicLayout } from './routes.testUtil';
 
 const setup = () => setupIntegrationTest('/');
